backend/requirements.txt
flask
flask-cors
flask-mail
groq
python-dotenv
psycopg2-binary
bcrypt
pyjwt
cloudinary
PyPDF2
pillow
langfuse